Set up CI run
Set up the (front-end, for the moment) to run on a CI server. This
involved hammering out a couple versioning issues, correcting a path
error in assignmentList spec, adding a Webpack loader for images in the
Karma config file, and adding the CI config file for CircleCI.

I should point out that it is fantastically unlikely that we will keep
using CircleCI, it just happens to be a CI server that I like, so it's
the one that I'm using for the moment.