BHA-97:  Implement SSL support, added tests.
BHA-96: Added a script for testing an SSL connection on url. Will alter url or ip as things change.

Fixed the url in sslTest.py to point towards current instance

Changed the url to point to the bha site where the app is going to be hosted.  Confirmed that they have their own SSL certificate implemented. Also edited test to work with Python 3.4