Revise django config to not use env vars in dev
This makes the the simplest possible change to fix how using
environment variables breaks in dev, which is that it not
just defaults to the dev settings. This probably isn't the
cleanest solution, but it does solve the problem. What I would
consider to be a better solution would be to have the environment
variables set on the VM, but that unfortunately uwsgi doesn't seem
to pick up environment variables in some cases. So this is the
shim for the moment.