BHA-112: Change infra tests regarding networking
Change the infrastructure tests regarding networking are organized.  In
particular, it adds some tests to make sure that UWSGI is installed and
has opened a unix socket to serve the Django app, and it adds a small
test suite to makes sure that the Nginx config file is in place and set
relatively correctly. And, also, a test to make sure that Nginx is
actually serving something over http port 80.