Remove email -> admin serializaion behavior
Remove the serialization behavior that would examine the emails of new
users and automatically make the administrators based on whether they
had an email from a particular domain. The intent is that a new user can
only be made staff as a result of the action of a current staff member.